Eout Analitico

+++ generatorPoints.jl
@@ -11,17 +11,17 @@ type Legendre
     this = new()
     this.qf = qf
     this.sigma = sigma
-    this.poly = [Float64(pol )for pol in sp.legendre(qf, monic=true)]    
+    this.poly = [pol for pol in sp.legendre(qf, monic=true)]
     this.generatorPoints =  function (n)
       xs = sort([ this.generateXPoint() for i=1:n])
       ys = [this.generateYPoint(x) for x in xs]
       return xs,ys
     end
     this.generateXPoint = function ()
-      return Float64(rand(Bool) ? rand() : -rand())
+      return rand(Bool) ? rand() : -rand()
     end
     this.generateYPoint = function (x)
-      return Float64(sum([this.poly[i]*(x^(i-1)) for i = 1:length(this.poly)]) + ( rand(Bool)? rand(0.0:0.001:this.sigma) : -rand(0.0:0.001:this.sigma)))
+      return sum([this.poly[i]*(x^(i-1)) for i = 1:length(this.poly)]) + ( rand(Bool)? rand(0.0:0.001:this.sigma) : -rand(0.0:0.001:this.sigma))
     end
 
     return this
